feat(assessment): Validations and externalizations

+++ assessment/src/main/java/com/cloudbees/assessment/application/service/train/TrainServiceImpl.java
@@ -1,6 +1,9 @@
 package com.cloudbees.assessment.application.service.train;
 
+import com.cloudbees.assessment.domain.dto.TrainDto;
 import com.cloudbees.assessment.domain.entity.Train;
+import com.cloudbees.assessment.domain.mapper.TrainMapper;
+import com.cloudbees.assessment.infrastructure.repository.TicketRepository;
 import com.cloudbees.assessment.infrastructure.repository.TrainRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,10 +13,11 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
-@RequiredArgsConstructor(onConstructor = @__({ @Autowired}))
+@RequiredArgsConstructor(onConstructor = @__({@Autowired}))
 public class TrainServiceImpl implements TrainService {
-    
+
     private final TrainRepository trainRepository;
+    private final TicketRepository ticketRepository;
 
     @Override
     public List<Train> getAllTrains() {
@@ -26,18 +30,23 @@ public class TrainServiceImpl implements TrainService {
     }
 
     @Override
-    public Train updateTrain(Train train) {
-        if(trainRepository.existsById(train.getId())) {
-            return trainRepository.save(train);
+    public Train updateTrain(Long id, TrainDto trainDto) {
+        var train = trainRepository.findById(id);
+        if (train.isPresent()) {
+            return trainRepository.save(TrainMapper.fromDto(id, trainDto, train.get()));
         } else {
-            throw new NoSuchElementException("Train with id " + train.getId() + " doesn't exist");
+            throw new NoSuchElementException("Train with id " + id + " doesn't exist");
         }
     }
 
     @Override
     public void deleteTrain(Long id) {
-        if(trainRepository.existsById(id)) {
-            trainRepository.deleteById(id);
+        if (trainRepository.existsById(id)) {
+            if (!ticketRepository.existsTicketByTrainId(id)) {
+                trainRepository.deleteById(id);
+            } else {
+                throw new NoSuchElementException("There is at least one ticket for user with id " + id);
+            }
         } else {
             throw new NoSuchElementException("Train with id " + id + " doesn't exist");
         }
