+++ assessment/src/main/java/com/cloudbees/assessment/application/service/user/UserServiceImpl.java
@@ -1,6 +1,9 @@
 package com.cloudbees.assessment.application.service.user;
 
+import com.cloudbees.assessment.domain.dto.UserDto;
 import com.cloudbees.assessment.domain.entity.User;
+import com.cloudbees.assessment.domain.mapper.UserMapper;
+import com.cloudbees.assessment.infrastructure.repository.TicketRepository;
 import com.cloudbees.assessment.infrastructure.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,13 +11,13 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor(onConstructor = @__({ @Autowired}))
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final TicketRepository ticketRepository;
 
     @Override
     public List<User> getAllUsers() {
@@ -27,18 +30,23 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User updateUser(User user) {
-        if(userRepository.existsById(user.getId())) {
-            return userRepository.save(user);
+    public User updateUser(Long id, UserDto userDto) {
+        var user = userRepository.findById(id);
+        if(user.isPresent()) {
+            return userRepository.save(UserMapper.fromDto(id, userDto, user.get()));
         } else {
-            throw new NoSuchElementException("User with id " + user.getId() + " doesn't exist");
+            throw new NoSuchElementException("User with id " + id + " doesn't exist");
         }
     }
 
     @Override
     public void deleteUser(Long id) {
         if(userRepository.existsById(id)) {
-            userRepository.deleteById(id);
+            if(!ticketRepository.existsTicketByUserId(id)) {
+                userRepository.deleteById(id);
+            } else {
+                throw new NoSuchElementException("There is at least one ticket for user with id " + id);
+            }
         } else {
             throw new NoSuchElementException("User with id " + id + " doesn't exist");
         }
